Move Reverse Proxy logs to the Reverse Proxy menu
